+++ app/Exceptions/Zip/ZipException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Exceptions\Zip;
+
+use RuntimeException;
+
+abstract class ZipException extends RuntimeException
+{
+}